+++ app/src/main/kotlin/org/fossify/gallery/helpers/Config.kt
@@ -217,9 +217,14 @@ class Config(context: Context) : BaseConfig(context) {
         get() = prefs.getBoolean(MUTE_VIDEOS, false)
         set(muteVideos) = prefs.edit().putBoolean(MUTE_VIDEOS, muteVideos).apply()
 
-    var openVideosOnSeparateScreen: Boolean
-        get() = prefs.getBoolean(OPEN_VIDEOS_ON_SEPARATE_SCREEN, false)
-        set(openVideosOnSeparateScreen) = prefs.edit().putBoolean(OPEN_VIDEOS_ON_SEPARATE_SCREEN, openVideosOnSeparateScreen).apply()
+    var gestureVideoPlayer: Boolean
+        get() = prefs.getBoolean(GESTURE_VIDEO_PLAYER, false)
+        set(gestureVideoPlayer) = prefs.edit().putBoolean(GESTURE_VIDEO_PLAYER, gestureVideoPlayer)
+            .apply()
+
+    var videoPlayerType: Int
+        get() = prefs.getInt(VIDEO_PLAYER_TYPE, VIDEO_PLAYER_APP)
+        set(videoPlayerType) = prefs.edit().putInt(VIDEO_PLAYER_TYPE, videoPlayerType).apply()
 
     var displayFileNames: Boolean
         get() = prefs.getBoolean(DISPLAY_FILE_NAMES, false)