+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added "On video tap" option to choose between in-app player or default player app ([#917])
+
 ### Fixed
 - Fixed unnecessary "Video has no audio" toast when looping videos ([#876])
 
@@ -292,6 +295,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#830]: https://github.com/FossifyOrg/Gallery/issues/830
 [#831]: https://github.com/FossifyOrg/Gallery/issues/831
 [#876]: https://github.com/FossifyOrg/Gallery/issues/876
+[#917]: https://github.com/FossifyOrg/Gallery/issues/917
 
 [Unreleased]: https://github.com/FossifyOrg/Gallery/compare/1.12.0...HEAD
 [1.12.0]: https://github.com/FossifyOrg/Gallery/compare/1.11.0...1.12.0